+++ rasa/nlu/test.py
@@ -19,7 +19,7 @@ from typing import (
 
 from rasa.constants import TEST_DATA_FILE, TRAIN_DATA_FILE
 from rasa.model import get_model
-from rasa.utils.io import create_path
+from rasa.utils.io import create_path, create_dir
 from rasa.nlu import config, training_data, utils
 from rasa.nlu.utils import write_to_file
 from rasa.nlu.components import ComponentBuilder
@@ -713,7 +713,7 @@ def run_evaluation(
     }  # type: Dict[Text, Optional[Dict]]
 
     if report:
-        utils.create_dir(report)
+        create_dir(report)
 
     intent_results, entity_results = get_eval_data(interpreter, test_data)
 
@@ -830,7 +830,7 @@ def cross_validate(
         nlu_config = config.load(nlu_config)
 
     if report:
-        utils.create_dir(report)
+        create_dir(report)
 
     trainer = Trainer(nlu_config)
     trainer.pipeline = remove_pretrained_extractors(trainer.pipeline)
